requirements.txt
spacy==3.2.1
st_annotated_text==3.0.0
streamlit==1.5.0
